refactor: separate panels into own components
