+++ zenith/src/lib/modules/launch-editor/launch-editor.svelte
@@ -12,32 +12,29 @@
 	export let onNodeChanged: $$Props['onNodeChanged'];
 </script>
 
-<div class="w-full h-full overflow-hidden">
-	<h1>Launch Editor</h1>
-	<Accordion.Root class="w-full">
-		{#each subsystems as subsystem}
-			<Accordion.Item value={subsystem.name}>
-				<Accordion.Trigger>{subsystem.name}</Accordion.Trigger>
-				<Accordion.Content>
-					<LazyLoadedList
-						ulClasses="overflow-x-hidden"
-						items={subsystem.packages.flatMap((pkg) =>
-							pkg.executables.map((exec) => ({ package: pkg.name, executable: exec }))
-						)}
-						itemsPerChunk={20}
-						let:item={node}
-					>
-						<li class="flex flex-wrap gap-4">
-							<Label>{`${node.package}::${node.executable}`}</Label>
-							<Switch
-								checked={activeNodes.some((n) => isNodeEqual(node, n))}
-								onCheckedChange={(checked) => onNodeChanged(node, checked)}
-								id={`${node.package}::${node.executable}`}
-							/>
-						</li>
-					</LazyLoadedList>
-				</Accordion.Content>
-			</Accordion.Item>
-		{/each}
-	</Accordion.Root>
-</div>
+<Accordion.Root class="w-full">
+	{#each subsystems as subsystem}
+		<Accordion.Item value={subsystem.name}>
+			<Accordion.Trigger>{subsystem.name}</Accordion.Trigger>
+			<Accordion.Content>
+				<LazyLoadedList
+					ulClasses="overflow-x-hidden"
+					items={subsystem.packages.flatMap((pkg) =>
+						pkg.executables.map((exec) => ({ package: pkg.name, executable: exec }))
+					)}
+					itemsPerChunk={20}
+					let:item={node}
+				>
+					<li class="flex flex-wrap gap-4">
+						<Label>{`${node.package}::${node.executable}`}</Label>
+						<Switch
+							checked={activeNodes.some((n) => isNodeEqual(node, n))}
+							onCheckedChange={(checked) => onNodeChanged(node, checked)}
+							id={`${node.package}::${node.executable}`}
+						/>
+					</li>
+				</LazyLoadedList>
+			</Accordion.Content>
+		</Accordion.Item>
+	{/each}
+</Accordion.Root>
